feat(count): prevent read file meta simultaneously; filter using bloom_filter
