+++ tskv/src/compaction/flush.rs
@@ -447,7 +447,7 @@ pub mod flush_tests {
     use std::str::FromStr;
     use std::sync::Arc;
 
-    use lru_cache::ShardedCache;
+    use lru_cache::asynchronous::ShardedCache;
     use memory_pool::{GreedyMemoryPool, MemoryPoolRef};
     use models::codec::Encoding;
     use models::schema::{ColumnType, TableColumn, TskvTableSchema};
